# Significantly improve the speed of shader compilation in compatibility backend