@Calinou 2024-01-24
I've benchmarked starting then exiting project manager with `--quit`. I've also tried the same with a 2D project featuring only a Sprite2D, but couldn't notice a measurable difference. I've removed the shader cache at `~/.cache/nvidia/GLCache/` between runs.

*Project manager always uses OpenGL unless otherwise specified on the command line.*

<details open>

<summary>PC specifications</summary>

- **CPU:** Intel Core i9-13900K
- **GPU:** NVIDIA GeForce RTX 4090
- **RAM:** 64 GB (2×32 GB DDR5-5800 C30)
- **SSD:** Solidigm P44 Pro 2 TB
- **OS:** Linux (Fedora 39)
</details>

#### Before

```
  Time (mean ± σ):      2.817 s ±  0.089 s    [User: 1.538 s, System: 0.227 s]
  Range (min … max):    2.703 s …  2.967 s    10 runs
```

#### After

```
  Time (mean ± σ):      2.720 s ±  0.323 s    [User: 1.542 s, System: 0.220 s]
  Range (min … max):    1.827 s …  2.969 s    10 runs
```