@clayjohn 2024-01-25
The project manager only uses one shader so I'm not surprised you aren't seeing much of a benefit. 

The MRP I used compiles 30 shaders in a frame which makes it a little easier to see the difference