@Alex2782 2024-01-26
[/gles3/shaders/scene.glsl](https://github.com/godotengine/godot/blob/5034478611697358f4e135c69f364ef6d8cd54dc/drivers/gles3/shaders/scene.glsl#L11) There are 5 `DISABLE_` defines in the file, can they also be set to `true` without hesitation?

On `moto e5 play` and `Huawei Mediapad T3 AGS-W09` (both Adreno (TM) 308) there is a strange `glLinkProgram` shader freeze bug. 2D scenes work as soon as the flags are set to `true`. 3D scenes unfortunately not yet.  

----------------------
I performed this measurement on an Android device, for `VERTEX` and `FRAGMENT`
It is a simple 2D/3D project. 4 vertex and 4 fragment shaders are compiled.
(DEBUG, Samsung Tab S7, no cache and no optimizations). 

```cpp
		uint64_t start = OS::get_singleton()->get_ticks_msec();
		glShaderSource(spec.vert_id, 1, &cstr, nullptr);
		glCompileShader(spec.vert_id);
		uint64_t end = OS::get_singleton()->get_ticks_msec();
		print_line("glCompileShader: ", (end-start), " msec, name: ", name, " VERTEX");
```

### CanvasShaderGLES3



- **VERTEX** is compiled up to 5x faster
- **FRAGMENT** is compiled up to 10x faster

<details><summary>Logs CanvasShaderGLES3</summary>

-----------------------------------------

**DISABLE_LIGHTING = true**
```
glCompileShader:  3  msec, name:  CanvasShaderGLES3  VERTEX
glCompileShader:  4  msec, name:  CanvasShaderGLES3  FRAGMENT
glCompileShader:  4  msec, name:  CanvasShaderGLES3  VERTEX
glCompileShader:  5  msec, name:  CanvasShaderGLES3  VERTEX
glCompileShader:  3  msec, name:  CanvasShaderGLES3  VERTEX
glCompileShader:  3  msec, name:  CanvasShaderGLES3  FRAGMENT
glCompileShader:  3  msec, name:  CanvasShaderGLES3  VERTEX
glCompileShader:  4  msec, name:  CanvasShaderGLES3  FRAGMENT
glCompileShader:  3  msec, name:  CanvasShaderGLES3  VERTEX
glCompileShader:  4  msec, name:  CanvasShaderGLES3  FRAGMENT
```

------------------

**DISABLE_LIGHTING = false** 
```
glCompileShader:  15  msec, name:  CanvasShaderGLES3  VERTEX
glCompileShader:  31  msec, name:  CanvasShaderGLES3  FRAGMENT
glCompileShader:  15  msec, name:  CanvasShaderGLES3  VERTEX
glCompileShader:  31  msec, name:  CanvasShaderGLES3  FRAGMENT
glCompileShader:  15  msec, name:  CanvasShaderGLES3  VERTEX
glCompileShader:  31  msec, name:  CanvasShaderGLES3  FRAGMENT
glCompileShader:  14  msec, name:  CanvasShaderGLES3  VERTEX
glCompileShader:  30  msec, name:  CanvasShaderGLES3  FRAGMENT
glCompileShader:  14  msec, name:  CanvasShaderGLES3  VERTEX
glCompileShader:  30  msec, name:  CanvasShaderGLES3  FRAGMENT
```
</details>

-----------------------------------------

### SceneShaderGLES3

 sometimes **up to 6x faster**, I have lowered a lot of settings, with default values it is perhaps even faster.

<details><summary>Logs SceneShaderGLES3 </summary>



-------------------------------------

**DISABLE_ * = true**
```
glCompileShader:  6  msec, name:  SceneShaderGLES3  VERTEX
glCompileShader:  8  msec, name:  SceneShaderGLES3  FRAGMENT
glCompileShader:  5  msec, name:  SceneShaderGLES3  VERTEX
glCompileShader:  8  msec, name:  SceneShaderGLES3  FRAGMENT
glCompileShader:  3  msec, name:  SceneShaderGLES3  VERTEX
glCompileShader:  4  msec, name:  SceneShaderGLES3  FRAGMENT
glCompileShader:  4  msec, name:  SceneShaderGLES3  VERTEX
glCompileShader:  5  msec, name:  SceneShaderGLES3  FRAGMENT
```

**DISABLE_ * = false**
```
glCompileShader:  9  msec, name:  SceneShaderGLES3  VERTEX
glCompileShader:  36  msec, name:  SceneShaderGLES3  FRAGMENT
glCompileShader:  7  msec, name:  SceneShaderGLES3  VERTEX
glCompileShader:  23  msec, name:  SceneShaderGLES3  FRAGMENT
glCompileShader:  5  msec, name:  SceneShaderGLES3  VERTEX
glCompileShader:  5  msec, name:  SceneShaderGLES3  FRAGMENT
glCompileShader:  5  msec, name:  SceneShaderGLES3  VERTEX
glCompileShader:  5  msec, name:  SceneShaderGLES3  FRAGMENT
```
</details>

![image](https://github.com/godotengine/godot/assets/41921395/344443b2-8ac9-4f4f-aa41-db56f90bede8)