@clayjohn 2024-01-26
> [/gles3/shaders/scene.glsl](https://github.com/godotengine/godot/blob/5034478611697358f4e135c69f364ef6d8cd54dc/drivers/gles3/shaders/scene.glsl#L11) There are 5 `DISABLE_` defines in the file, can they also be set to `true` without hesitation?

Potentially, yes. I would like to explore the situation for scene.glsl a little bit more before committing to a decision. The tradeoffs are slightly different than with canvas.glsl. 

The biggest different is that a typical 3D scene is more likely to actually use multiple render modes whereas in canvas.glsl, users often only need one of the render modes. So saving the compile cost for the unused render modes is significant. 

The other difference is that in 3D, lighting is much more likely to be used. So if we disable lights by default, then we compile two things, the unlit version (at load time) and the lit version at run time. Since the lit version is so much more expensive to compile we are more likely to notice stutters during gameplay.