@akien-mga 2024-03-10
Cherry-picked for 4.2.2.
Cherry-picked for 4.1.4.